// SolitaireLib/Games/ClockEngine/Model/IGame.ts
import { IGameBase } from "~CardLib/Model/IGameBase";
import { IPile } from "~CardLib/Model/IPile";
import { GameOptions } from "./GameOptions";

export interface IGame extends IGameBase {
    readonly options: GameOptions;
    readonly foundations: IPile[];
    readonly tableaux: IPile[];
    readonly stock?: IPile;
    readonly waste?: IPile;
}